+++ History.md
@@ -96,6 +96,12 @@
   HTTP.call("GET", url, ...);
   ```
 
+* The log-suppressing flags on errors in `ddp-client` and `ddp-server` have been
+  changed from `expected` to `_expectedByTest` in order to avoid inadvertently
+  silencing errors in production.
+  [Issue #6912](https://github.com/meteor/meteor/issues/6912)
+  [PR #9515](https://github.com/meteor/meteor/pull/9515)
+
 * [`cordova-lib`](https://github.com/apache/cordova-cli) has been updated to
   version 7.1.0, [`cordova-android`](https://github.com/apache/cordova-android/)
   has been updated to version 6.3.0, and [`cordova-ios`](https://github.com/apache/cordova-ios/)