+++ History.md
@@ -194,6 +194,7 @@
 
 * Mongo has been upgraded to version 3.6.2 for 64-bit OS', and 3.2.18 for
   32-bit OS'.
+  [PR #9533](https://github.com/meteor/meteor/pull/9533)
 
 ## v1.6.0.1, 2017-12-08