+++ History.md
@@ -6,7 +6,7 @@
   [8.9.4](https://nodejs.org/en/blog/release/v8.9.4/).
 
 * The `meteor-babel` npm package (along with its Babel-related
-  dependencies) has been updated to version 7.0.0-beta.37-1, a major
+  dependencies) has been updated to version 7.0.0-beta.38, a major
   update from Babel 6. Thanks to the strong abstraction of the
   `meteor-babel` package, the most noticeable consequence of the Babel 7
   upgrade is that the `babel-runtime` npm package has been replaced by